==> nb/2.export-to-manually-annotate.ipynb <==
import sqlite3

import polars as pl

with sqlite3.connect("../data/project.db") as conn:
    labels_df = pl.read_database("SELECT * from druglabel;", conn)

manual_df = (
    labels_df
    .filter(pl.col("id").le(101) & pl.col("id").gt(1))
    .select("id", "set_id", "label_id", "spl_version", "title", "CO", "DI")
    .melt(id_vars=["set_id", "label_id", "spl_version", "title"], value_vars=["CO", "DI"], 
          variable_name="section", value_name="label")
)

manual_df

(
    manual_df
    .write_excel("../data/manual_annotations.xlsx")
)
