==> nb/3.build-consensus.ipynb <==
import pandas as pd
import polars as pl

pre_annotated_df = pl.read_csv("../data/to_manually_annotate.csv")
uog_df = pl.read_csv("../data/undina_annotations.tsv", separator="\t")
mnz_df = pl.read_csv("../data/michael_annotations.csv")

mnz_df = (
    pl.concat([
        pre_annotated_df,
        mnz_df.rename({"title": "_title", "text": "_text"})
    ], how="horizontal")
)
assert len(mnz_df.filter(pl.col("set_id").ne(pl.col("_title")))) == 0
mnz_df = (
    mnz_df
    .select("set_id", "label_id", "spl_version", "title", "section", "label", "annotation")
)

merged_df = (
    uog_df
    .join(mnz_df, on=["set_id", "label_id", "spl_version", "section"])
    .filter(pl.col("section").eq("DI"))
    .with_columns(pl.col("label", "title").str.replace_all("\n", " "))
    .select("set_id", "label_id", "spl_version", "title", "section", "label", "uog_annotation", mnz_annotation = "annotation")
)

merged_df.head(2)

diff_df = (
    merged_df
    .with_columns(
        pl.col("uog_annotation", "mnz_annotation")
            .str.split(",")
            .list.eval(
                pl.element()
                .str.strip_chars()
                .str.to_lowercase()
                .str.replace(".", "", literal=True)
                .str.replace("-", " ")
            )
            .list.unique()
    )
    .with_columns(
        uog_diff=pl.col("uog_annotation").list.set_difference(pl.col("mnz_annotation")),
        mnz_diff=pl.col("mnz_annotation").list.set_difference(pl.col("uog_annotation")),
    )
    # Replace empty list with null
    .with_columns(
        uog_diff=pl.when(pl.col("uog_diff").list.len() > 0).then(pl.col("uog_diff")),
        mnz_diff=pl.when(pl.col("mnz_diff").list.len() > 0).then(pl.col("mnz_diff")),
    )
    .with_columns(consensus=pl.col("uog_annotation").list.set_intersection(pl.col("mnz_annotation")))
)

diff_df.head(2)

(
    diff_df
    .with_columns(pl.col(pl.List(pl.Utf8)).list.join(", "))
    .write_csv("../data/initial_diff.csv")
)

(
    diff_df
    .filter(pl.col("uog_diff").is_not_null() | pl.col("mnz_diff").is_not_null())
    .with_columns(pl.col(pl.List(pl.Utf8)).list.join(","))
)
