==> nb/4.get_embeddings.ipynb <==
import pandas as pd
import polars as pl
from sqlmodel import Session, select
from sentence_transformers import SentenceTransformer

from undina_llm.db import SessionManager
from undina_llm.models import Response, Prompt, SystemPrompt, DrugLabel

# 1. Load the annotations

manual_df = (
    pl.read_csv("../data/merged_diff.csv")
    .with_columns(
        pl.col("consensus", "additions")
        .str.split(",")
        .list.eval(pl.element().str.strip_chars())
        .list.unique()
    )
    .with_columns(
        manual_annotation=(
            pl.col("consensus")
            .list.set_union(pl.col("additions"))
            .list.unique()
        )
    )
    .select("set_id", "label_id", "spl_version", "title", "section", "label", "manual_annotation")
)

manual_df.with_columns(pl.col("manual_annotation").list.join(", ")).write_csv("../data/manual_annotations.csv")

manual_df.head(2)

SQLITE_FILE = "sqlite:///../data/project.db"

manager = SessionManager(SQLITE_FILE)

with Session(manager.engine) as session:
    responses_df = pl.DataFrame(session.exec(select(Response).where((Response.section == "DI") & (Response.prompt_id == 2))).all())
    system_prompts_df = pl.DataFrame(session.exec(select(SystemPrompt)).all())
    prompts_df = pl.DataFrame(session.exec(select(Prompt)).all())
    drug_labels_df = pl.DataFrame(session.exec(select(DrugLabel)).all())

responses_df.head(2)

gpt_df = (
    responses_df
    .join(drug_labels_df, left_on=["drug_label_id"], right_on=["id"])
    .drop(["system_prompt_id", "prompt_id", "drug_label_id", "section", "model", "temperature"])
    .select(
        "set_id", "label_id", "spl_version", "title", pl.lit("DI").alias("section"), 
        pl.col("DI").alias("label"), pl.col("response").alias("gpt_annotation")
    )
)

gpt_df.write_csv("../data/gpt_annotations.csv")

gpt_df.head(2)

# 2. Gather the unique annotation strings

manual_annotations = (
    manual_df
    .explode("manual_annotation")
    .select("manual_annotation")
    .drop_nulls()
    .unique()
    ["manual_annotation"]
    .to_list()
)
print(len(manual_annotations))

gpt_annotations = (
    gpt_df
    .select(pl.col("gpt_annotation").str.split(","))
    .explode("gpt_annotation")
    .filter(pl.col("gpt_annotation").ne('""'))
    .select(
        pl.col("gpt_annotation")
        .str.strip_chars()
        .str.to_lowercase()
    )
    .unique()
    ["gpt_annotation"]
    .to_list()
)

print(len(gpt_annotations))

# 3. Gather RxNorm ingredient names

rxnorm_df = (
    pl.read_csv("../data/rxnorm_ingredients_synonyms.tsv", separator="\t")
    .melt(id_vars=["concept_id"], value_vars=["concept_name", "concept_synonym_name"], value_name="name")
    .drop("variable")
    .drop_nulls("name")
)

rxnorm_annotations = rxnorm_df.unique("name")["name"].to_list()
print(len(rxnorm_annotations))

rxnorm_df.head(2)

# 4. Compute embeddings for each annotation

embed_model = SentenceTransformer('llmrails/ember-v1')

manual_embeddings = embed_model.encode(manual_annotations)
gpt_embeddings = embed_model.encode(gpt_annotations)
rxnorm_embeddings = embed_model.encode(rxnorm_annotations)

manual_embeddings_df = pd.DataFrame(manual_embeddings, index=manual_annotations)
gpt_embeddings_df = pd.DataFrame(gpt_embeddings, index=gpt_annotations)
rxnorm_embeddings_df = pd.DataFrame(rxnorm_embeddings, index=rxnorm_annotations)

manual_embeddings_df.to_csv("../data/manual_embeddings.csv")
gpt_embeddings_df.to_csv("../data/gpt_embeddings.csv")
rxnorm_embeddings_df.to_csv("../data/rxnorm_embeddings.csv.zst")
