==> nb/5.compare-embeddings.ipynb <==
import numpy as np
import pandas as pd
import polars as pl
import sklearn.metrics.pairwise

manual_embeddings_df = pd.read_csv("../data/manual_embeddings.csv", index_col=0)
gpt_embeddings_df = pd.read_csv("../data/gpt_embeddings.csv", index_col=0)
rxnorm_embeddings_df = pd.read_csv("../data/rxnorm_embeddings.csv.zst", index_col=0)

def compute_cosine_df(**kwargs):
    assert len(kwargs) == 2
    x_name, y_name = list(kwargs.keys())
    X = kwargs[x_name]
    Y = kwargs[y_name]
    
    return (
        pd.DataFrame(
            sklearn.metrics.pairwise.cosine_similarity(X, Y),
            index=X.index,
            columns=Y.index,
        )
        .melt(ignore_index=False, var_name=y_name, value_name="cosine_similarity")
        .rename_axis(index=x_name)
        .reset_index()
        .pipe(pl.DataFrame)
    )

# 1. Direct comparison between GPT and manual annotations using semantic similarity matching

DIRECT_MATCH_CUTOFF = 0.7097613

manual_vs_gpt_cosine_df = compute_cosine_df(manual_annotation=manual_embeddings_df, gpt_annotation=gpt_embeddings_df)

manual_annotations_df = pl.read_csv("../data/manual_annotations.csv")
gpt_annotations_df = pl.read_csv("../data/gpt_annotations.csv")

manual_df = (
    manual_annotations_df
    .drop_nulls("manual_annotation")
    .select(
        "set_id", "label_id", 
        (
            pl.col("manual_annotation")
            .str.split(",")
            .list.eval(pl.element().str.strip_chars())
        )
    )
    .explode("manual_annotation")
)

gpt_df = (
    gpt_annotations_df
    .drop_nulls("gpt_annotation")
    .select(
        "set_id", "label_id", 
        (
            pl.col("gpt_annotation")
            .str.split(",")
            .list.eval(pl.element().str.strip_chars().str.to_lowercase())
        )
    )
    .explode("gpt_annotation")
    .filter(pl.col("gpt_annotation").ne('""') & pl.col("gpt_annotation").str.len_chars().ne(0))
)

unmatched_rows = (
    manual_df
    .join(gpt_df, on=["set_id", "label_id"])
    .join(manual_vs_gpt_cosine_df, on=["manual_annotation", "gpt_annotation"], how="anti")
    .shape[0]
)
assert unmatched_rows == 0

merged_df = (
    manual_df
    .join(gpt_df, on=["set_id", "label_id"], how="full")
    .join(manual_vs_gpt_cosine_df, on=["manual_annotation", "gpt_annotation"])
)

merged_df.head(2)

manual_annotations_best_match_df = (
    merged_df
    .group_by("set_id", "label_id", "manual_annotation")
    .agg(pl.col("cosine_similarity").max())
    .with_columns(match=pl.col("cosine_similarity").ge(DIRECT_MATCH_CUTOFF))
)

gpt_annotations_best_match_df = (
    merged_df
    .group_by("set_id", "label_id", "gpt_annotation")
    .agg(pl.col("cosine_similarity").max())
    .with_columns(match=pl.col("cosine_similarity").ge(DIRECT_MATCH_CUTOFF))
)

(
    manual_annotations_best_match_df
    .select(
        N=pl.len(),
        P=pl.col("match").sum(),
    )
    .with_columns(recall=pl.col("P").truediv(pl.col("N")))
)

(
    gpt_annotations_best_match_df
    .select(
        N=pl.len(),
        P=pl.col("match").sum(),
    )
    .with_columns(precision=pl.col("P").truediv(pl.col("N")))
)

# 2. Evaluation using RxNorm semantic similarity. Best match only!

RXNORM_MATCH_CUTOFF = 0.9

rxnorm_df = (
    pl.read_csv("../data/rxnorm_ingredients_synonyms.tsv", separator="\t")
    .melt(id_vars=["concept_id"], value_vars=["concept_name", "concept_synonym_name"], value_name="name")
    .drop("variable")
    .drop_nulls("name")
)

manual_cosine_df = compute_cosine_df(manual=manual_embeddings_df, rxnorm=rxnorm_embeddings_df)
gpt_cosine_df = compute_cosine_df(gpt=gpt_embeddings_df, rxnorm=rxnorm_embeddings_df)

manual_eval_df = (
    manual_cosine_df
    .filter(pl.col("cosine_similarity").ge(RXNORM_MATCH_CUTOFF))
    .sort("cosine_similarity")
    .filter(pl.col("cosine_similarity").eq(pl.col("cosine_similarity").max().over("manual")))
    .join(rxnorm_df, left_on=["rxnorm"], right_on=["name"])
)

gpt_eval_df = (
    gpt_cosine_df
    .filter(pl.col("cosine_similarity").ge(RXNORM_MATCH_CUTOFF))
    .sort("cosine_similarity")
    .filter(pl.col("cosine_similarity").eq(pl.col("cosine_similarity").max().over("gpt")))
    .join(rxnorm_df, left_on=["rxnorm"], right_on=["name"])
)

manual_rxnorm_df = (
    manual_annotations_df
    .with_columns(
        pl.col("manual_annotation").str.split(",").list.eval(pl.element().str.strip_chars())
    )
    .explode("manual_annotation")
    .join(manual_eval_df, left_on=["manual_annotation"], right_on=["manual"])
    .select("set_id", "label_id", "concept_id")
)

gpt_rxnorm_df = (
    gpt_annotations_df
    .with_columns(
        pl.col("gpt_annotation").str.split(",").list.eval(pl.element().str.strip_chars())
    )
    .explode("gpt_annotation")
    .join(gpt_eval_df, left_on=["gpt_annotation"], right_on=["gpt"])
    .select("set_id", "label_id", "concept_id")
)

(
    manual_rxnorm_df
    .join(gpt_rxnorm_df, on=["set_id", "label_id"], suffix="_gpt", how="full")
    .with_columns(match=pl.col("concept_id").eq(pl.col("concept_id_gpt")))
    .group_by("set_id", "label_id", "concept_id")
    .agg(pl.col("match").any().cast(pl.UInt32))
    .select(
        N=pl.len(),
        P=pl.col("match").sum(),
    )
    .with_columns(recall=pl.col("P").truediv(pl.col("N")))
)

(
    gpt_rxnorm_df
    .join(manual_rxnorm_df, on=["set_id", "label_id"], suffix="_manual")
    .with_columns(match=pl.col("concept_id").eq(pl.col("concept_id_manual")))
    .group_by("set_id", "label_id", "concept_id")
    .agg(pl.col("match").any().cast(pl.UInt32))
    .select(
        N=pl.len(),
        P=pl.col("match").sum(),
    )
    .with_columns(precision=pl.col("P").truediv(pl.col("N")))
)
